feat(ember): public-safe ember_public package for the postgres demo
